+++ lib/renderer/src/renderer/gl/shaders/outline.frag
@@ -8,7 +8,7 @@ layout ( std140, binding = 2 ) uniform Uniforms
 {	
 	vec4 color;
 	float sensivity;
-	float thickness;	
+	uint thickness;	
 } uniforms;
 
 // Out.
